Add pub to methods
Non-pub methods can only be called from within the same file where they're defined

+++ src/main/java/ast/ir/instruction/Instruction.java
@@ -10,6 +10,6 @@ public interface Instruction {
     void accept(MethodVisitor jvm) throws CompilationException;
 
     //Get the cost for this operation (not counting inner code blocks)
-    int cost();
+    long cost();
 
 }
